+++ core/src/main/scala/me/shadaj/slinky/core/Component.scala
@@ -42,3 +42,8 @@ abstract class Component {
 
   def render(): ReactElement
 }
+
+abstract class StatelessComponent extends Component {
+  type State = Unit
+  def initialState: State = ()
+}